consolidate scripts
