+++ engine/pyproject.toml
@@ -8,11 +8,14 @@ license = "GNU General Public License v3.0"
 [tool.poetry.dependencies]
 python = ">=3.9,<3.10"
 pytest = "^6.2.4"
+web3 = "^5.25.0"
+typer = "^0.3.2"
 
 [tool.poetry.dev-dependencies]
-typer = "^0.3.2"
-uncompyle6 = "^3.7.4"
-pandas = "^1.3.3"
+pandas = "^1.3.5"
+
+[tool.poetry.scripts]
+mlth = { reference="bin/mlth.py", type="file" }
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
